+++ packages/frogbot/src/exports/search.ts
@@ -1,2 +1,3 @@
 export { postgresSearchAdapter } from '../search/drizzle/postgres/index.js';
+export { sqliteD1SearchAdapter } from '../search/drizzle/sqlite/d1/index.js';
 export { sqliteSearchAdapter } from '../search/drizzle/sqlite/index.js';